[ep2 cuda eemm] Allow for setting madgraph4gpu config from outside.
In order to be able to set e.g. -DMGONGPU_CURAND_ONHOST etc from
outside, they must not be overwritten in mgOnGpuConfig.h
Therefore, the section that sets one of those macros is now protected
such that it only has an effect if none of these macros is set.

Ported to epoch2 from b52f8f23